Add functions to add claim partition
By getting an input of two different locations alongside the associated
claim, a claim partition can be added. This gets added to two lists, one
flat list and another that associates chunks with claims. By getting all
the chunks that the claim encompasses, it ensures that every chunk that
the claim encompasses is recorded in order to find the claim easier.