Create custom claim tool
This is an object that holds the data required to recreate the claim
tool. Currently, the tool is assigned to a stick with a custom name,
though the real identifier is with the custom model data. This allows a
check for knowing if it's the correct item and to allow custom resource
packs to override the texture.