Complete claim overlap check on claim creation
This implementation takes the chunks that the claim takes up and gets a
list of preexisting claims that exist in these chunks. It then does a
rectangle overlap check based on the position of the existing claim
partition compared to the new claim partition and only continues if no
overlaps exist.