Replace Owner classes with PlayerState
To hold all kinds of data related to players on the server, a
PlayerState object has been created with a PlayerContainer to house
them. This replaces the currently existing but unused Owner storage as
that was originally intended to just house global claim permissions for
players. PlayerState will hold additional data such as claim limits.