Add PlayerState table to database
This table complements the PlayerState object to store data
persistently. It omits keys for the claims and permissions as those can
piggyback off previous table data, and only claim limits have to be
stored as of now. A function has been added to fetch from this table, as
well as a function specifically for getting claims by player for this
specific functionality.