Add functions to add and remove playerState from db
These functions take in a PlayerState object and modify the data to add
or remove player state data. Unlike with other database manipulation
functions that take in individual variables, these take in the whole
object to reduce the amount of extra code each time these functions need
to be called. Other functions will take on this change as well at a
later date.