Add main partition to claim
The main partition will be utilised to make it easier to locate claims
on claimlist. The centerpoint of the main partition will be utilised for
both chat info and teleporting in future implementations. Currently, the
first partition is assigned as the main partition, though manual
changing will be implemented later.