Change method of fetching claims from database
The original method fetched claims for the claimcontainer and claims for
the playerstate individually, leading to duplicate claim data in memory
and ultimately doubling how much memory was required alongside having to
manipulate the data in two places at once. This change fetches claims
based on players, using the same claim and partition data for both
playerstates and claimcontainer.