Prevent players from throwing out claim tool
To remove the player's ability from taking the claim tool out of their
inventory, a number of measures have taken place. They cannot move it to
another inventory by any means, nor put it in an item frame. Dropping
the claim tool is simply an easy way to get rid of it.