Add creation time data to claims
Creation date now has to be added to both the claim database and object
on creation, which is displayed upon running the claim info command.
This makes use of the Instant object to get the time in UTC, a timezone
that can be used to easily format into other timezones.