Add new Position and Area objects to define claims
Rather than making use of an integer pair to define a position value,
the Position object explicitly defines an x and z coordinate. This
reduces confusion in using the first and second keyword of a Pair object
to get coordinates.

The Area object reduces the amount of duplicate code that is generated
in the partition, builder, and resizer objects. Rather than having to
store the position values individually, this handles the storage of both
position values to define an area, with functions that handle fetching
position data moved to this class.