Separate Partition references from Claim container
The Partition object now removes the direct object reference to claim,
removing the circular dependency in having to fetch partitions with an
associated claim. This means that each function is now in reference to
only the partition object itself, rather than making use of world or
claim reference to get results for certain functions.

Partitions have been moved to their own container, aptly named
Repository. This new name will be the basis for future refactorings, as
it better suits the name of the pattern it is based on. Functions in
this are suited to only that of Partitions, usually involving their
space in the world.