Refactor PlayerAccess and create associated mapper
This mapper will be utilised by repositories such as claims and players
in order to fetch association data for player permissions in certain
claims. The PlayerAccess no longer contains a list of permissions, so
PlayerAccess objects themselves must be a list in itself. This allows
for a greater degree of flexibility when needing to retrieve association
data from multiple sources.