Enact major project restructuring
There are so many changes that had to be done that I can no longer be
bothered to break it up into more managable smaller commits, since many
of the changes broke the code in many other places, requiring almost all
the code to be changed.

Changes had mostly to do with giving each object its on single purpose
rather than bundling in all the database calls all in one place. This
required the creation of many repositories and a complete restructuring
of how claim data is fetched.