Add object and listener to move claims
The "delete claim" menu item has been replaced by a "move claim" button,
which gives the player a bell item that allows them to move the claim
upon placing it down. The position the player is allowed to place it in
has to be within the bounds of the claim area, and upon successfully
selecting a location, the old bell block disappears.