Fix tree claim griefing
This is a very situational issue, but all issues must be fixed. The best
way to solve this was to only allow tree growth for saplings placed
outside of a claim if the resulting tree would not result in generated
blocks ending up inside the claim. This sets up its own limitation but
prevents tree griefing entirely.