Change position system to fix inheriting problems
There's a fair bit of confusion with when to use position, position2d,
worldposition, and how they can be used. Position should now be the one
that is used whenever it doesn't matter whether it is 2d or 3d.
Position3d should be used when a 3D position is explicitly required.
WorldPosition now contains an instance of position rather than
explicitly having 2d coordinates stored. It's recommended that a 2d
position is used for WorldPosition though.