Reorganise further into more defined categories
This is an attempt to integrate the hexagonal architecture in a way that
suits this project.

API represents the inbound connection with the program. That is,
anything that wishes to interface with the application must do so from
the given API endpoint classes. This should solely consist of interfaces
without any concrete implementations.

Domain represents the data the program is working with. This should
contain information about the data being stored, as well as how the data
is expected to be persisted. Data objects can contain whatever they want
so long as it pertains only to itself, while repositories should only be
an interface to a potential data store.

Infrastructure is the concrete implementations of the inbound and
outbound interfaces. It should describe how the data is manipulated as
well as the types of storage persistence.

Interaction is simply the user facing ways to interact with the
application. This consists of things like commands the player can type
as well as listeners that perform reactions on certain player actions.

Utils is a miscellaneous folder, notably used for functions that don't
belong anywhere else. Might change in the future.