Fix placing fluids through entities
Interesting bug occurred when an entity is blocking the block you want
to place fluids on. Doing so allows you to place the fluid regardless of
any protection applied. Turns out the stupid easy fix was to just use
the PlayerBucketEmpty event that I didn't know existed instead of the
generic interact event.