Overhaul claim permission handler
Multiple changes were done as was required to even begin to test this
implementation. This removes the permission parenting system, changes
the location check to obtain a list of locations instead of just one,
and loops through each location to gather all affected claims.

This was all to fix the issue of multi-block placement and also
streamlines the existing codebase.