Change localization provider to separate console locale
As per player locale will be a feature added later down the line, the
localization provider will always require the player's locale to move
the decision making process from the presentation to the infrastructure
layer. This also solves the current issue where it assumes the first
string is a locale when it isn't in language keys that require
arguments to be set.